Update: Interface refactoring
